Data model defined
The data model for the application will be two objects - a ```pl.kmejka.reminders.model.ReminderRequest``` which will be the input to the application. The input has to contain only the title of the needed reminder. The actual reminder will be stored in the ```pl.kmejka.reminders.model.Reminder``` object which will be created basing on the title from the request, a ```java.util.UUID``` id, the user id taken from the path, the start and end date, and the channel which will taken from the LGI api.
The data objects will be serializable to json. The serialization is available immediatly, since the dropwizard-scala library contains an already initialised jetty module which contains a scala jackson mapper.