Add implementation of saving reminders
The reminders are now stored by the ```RemindersDAO``` object in a ConcurrentHashMap (I failed to use native ```scala.collection.concurrent.Map```). The UUID generation is now done in the DAO object. The path to the created resource, along with the object itself, is now returned by the application. Changed the static constructor of the ```Reminder``` to use only the title and the userId (the rest will be filled in other places, so we save defining some nulls in the object creation).