Add implementation of getting reminders
Added implementation of ```getReminder``` method which returns the reminder for a given ```reminderId: String```. Other methods of the dao object are now waiting for implementation, for now they only return empty Lists.
The two existing methods in the DAO class are now tested using the ```scalatest``` library with a junit runner.
In the future the implementation of DAO needs to be changed to use the concurrent map from scala library, to use a more functional and monadic approach.